Rename ingestibles.py to manifest.py to reflect its purpose
