+++ src/beneficiations/abstract_custom_beneficiation.py
@@ -13,7 +13,7 @@ import logging
 from abc import ABC, abstractmethod
 from typing import Any
 
-from src.extraction.ingestibles import IngestibleDataclasses
+from src.extraction.manifest import IngestibleDataclasses
 
 # ---Constants
 logger = logging.getLogger(__name__)
