+++ src/beneficiations/parsers/json_parser.py
@@ -15,7 +15,7 @@ from typing import Any, Type, TypeVar
 
 from src.beneficiations.parsers.parser import Parser
 from src.blueprints import RawDatafile, RawDataset, RawExperiment, RawProject
-from src.extraction.ingestibles import IngestibleDataclasses
+from src.extraction.manifest import IngestibleDataclasses
 from src.profiles import profile_consts as pc
 
 # ---Constants
