+++ src/beneficiations/parsers/parser.py
@@ -9,7 +9,7 @@ import logging
 from abc import ABC, abstractmethod
 from typing import Any, Dict
 
-from src.extraction.ingestibles import IngestibleDataclasses
+from src.extraction.manifest import IngestibleDataclasses
 
 # ---Constants
 logger = logging.getLogger(__name__)
