+++ src/extraction/manifest.py
