+++ src/extraction/metadata_extractor.py
@@ -4,7 +4,7 @@
 from abc import ABC, abstractmethod
 from pathlib import Path
 
-from src.extraction.ingestibles import IngestibleDataclasses
+from src.extraction.manifest import IngestibleDataclasses
 
 
 class IMetadataExtractor(ABC):
