+++ src/profiles/idw/custom_beneficiation.py
@@ -26,7 +26,7 @@ from src.beneficiations.abstract_custom_beneficiation import AbstractCustomBenef
 from src.blueprints import RawDatafile, RawDataset, RawExperiment, RawProject
 from src.blueprints.common_models import GroupACL, UserACL
 from src.blueprints.custom_data_types import Username
-from src.extraction.ingestibles import IngestibleDataclasses
+from src.extraction.manifest import IngestibleDataclasses
 from src.miners.utils import datafile_metadata_helpers
 from src.mytardis_client.enumerators import DataClassification
