+++ src/profiles/ro_crate/ro_crate_parser.py
@@ -2,7 +2,6 @@
 Classes for handling an RO-crate input into ingestible data objects
 """
 
-
 import json
 
 # ---Imports
@@ -23,7 +22,7 @@ from src.blueprints.datafile import RawDatafile  # pylint: disable=duplicate-cod
 from src.blueprints.dataset import RawDataset  # pylint: disable=duplicate-code
 from src.blueprints.experiment import RawExperiment  # pylint: disable=duplicate-code
 from src.blueprints.project import RawProject  # pylint: disable=duplicate-code
-from src.extraction.ingestibles import IngestibleDataclasses
+from src.extraction.manifest import IngestibleDataclasses
 from src.extraction.metadata_extractor import (  # pylint: disable=duplicate-code
     IMetadataExtractor,
 )
